Refactor _socket2/_socket3 to move more truly common methods into SocketMixin.
