+++ docs/api/gevent._socket2.rst
@@ -2,5 +2,10 @@
  :mod:`gevent._socket2` -- Python 2 socket module
 ==================================================
 
+.. caution:: Some of the docstrings are automatically generated and
+             may be correct for Python 3 but not Python 2. Please see
+             the standard library documentation.
+
 .. automodule:: gevent._socket2
     :members:
+    :inherited-members:
