+++ docs/api/gevent._socket3.rst
@@ -4,3 +4,4 @@
 
 .. automodule:: gevent._socket3
     :members:
+    :inherited-members:
