+++ docs/api/gevent.socket.rst
@@ -52,6 +52,8 @@ Python 2 and Python 3, respectively.
        standard library sockets never have.
 
 
+
+
 .. toctree::
 
    Python 3 interface <gevent._socket3>
