+++ docs/changes/1724.bugfix
@@ -2,3 +2,6 @@ Remove the ``__dict__`` attribute from `gevent.socket.socket` objects. The
 standard library socket do not have a ``__dict__``.
 
 Noticed by Carson Ip.
+
+As part of this refactoring, share more common socket code between Python 2
+and Python 3.
